+++ src/__test__/dot.test.ts
@@ -1,0 +1,10 @@
+import { Digraph } from 'ts-graphviz';
+import { dot } from '../dot';
+
+describe('dot template tag', () => {
+  test('simple digraph', () => {
+    const G = dot`digraph {}`;
+    expect(G).toBeInstanceOf(Digraph);
+    expect(G.strict).toStrictEqual(false);
+  });
+});
