+++ src/__test__/parse.spec.ts
@@ -1,17 +1,9 @@
 import { Digraph } from 'ts-graphviz';
-import { dot, parse } from '../usecase';
+import { parse } from '../parse';
 
 describe('parse function', () => {
   test('simple digraph', () => {
     const G = parse('digraph {}');
-    expect(G).toBeInstanceOf(Digraph);
-    expect(G.strict).toStrictEqual(false);
-  });
-});
-
-describe('dot template tag', () => {
-  test('simple digraph', () => {
-    const G = dot`digraph {}`;
     expect(G).toBeInstanceOf(Digraph);
     expect(G.strict).toStrictEqual(false);
   });
