+++ src/index.ts
@@ -1,3 +1,4 @@
 export { SyntaxError } from './dot.pegjs';
 export { AST } from './ast';
-export { parse, dot } from './usecase';
+export { parse } from './parse';
+export { dot } from './dot';
